requirements.txt
pycryptodome>=3.10.1
json5>=0.9.5